+++ common/src/main/java/moe/qingu/geocraft/geography/fluidphysics/updater/FluidTaskQueue.java
@@ -41,7 +41,7 @@ public abstract class FluidTaskQueue {
 
     public abstract int size();
 
-    public abstract void queue(final int cx,final int cy,final int cz,final short taskID);
+    public abstract void queue(final int cx,final int cy,final int cz,final int taskID);
 
     public void queue(final int task){
         queue((task>>>4)&0xF,task>>>24,task & 0xF, (short) ((task>>>8)&0xFFFF));
